web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["nohkwanok <cloq@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
jinja2 = "^3.1.3"
black = "^24.2.0"
ruff = "^0.3.0"
gunicorn = "^21.2.0"
django-environ = "^0.11.2"
requests = "^2.31.0"
psycopg = "^3.1.18"
asyncpg = "^0.29.0"
aiohttp = "^3.9.3"
markdown = "^3.5.2"
django-filter = "^24.1"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.3.0"
django-allauth = "^0.61.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
